Remove duplicated context files
