+++ src/common/entity/context/get_entity_context.ts
@@ -1,6 +1,11 @@
+import type { HassEntity } from "home-assistant-js-websocket";
 import type { AreaRegistryEntry } from "../../../data/area_registry";
 import type { DeviceRegistryEntry } from "../../../data/device_registry";
-import type { EntityRegistryDisplayEntry } from "../../../data/entity_registry";
+import type {
+  EntityRegistryDisplayEntry,
+  EntityRegistryEntry,
+  ExtEntityRegistryEntry,
+} from "../../../data/entity_registry";
 import type { FloorRegistryEntry } from "../../../data/floor_registry";
 import type { HomeAssistant } from "../../../types";
 
@@ -11,27 +16,15 @@ interface EntityContext {
   floor: FloorRegistryEntry | null;
 }
 
-/**
- * Retrieves the context of an entity, including its associated device, area, and floor.
- *
- * @param entityId - The unique identifier of the entity to retrieve the context for.
- * @param hass - The Home Assistant object containing the registry data for entities, devices, areas, and floors.
- * @returns An object containing the entity, its associated device, area, and floor, or `null` for each if not found.
- *
- * The returned `EntityContext` object includes:
- * - `entity`: The entity registry entry, or `null` if the entity is not found.
- * - `device`: The device registry entry associated with the entity, or `null` if not found.
- * - `area`: The area registry entry associated with the entity or device, or `null` if not found.
- * - `floor`: The floor registry entry associated with the area, or `null` if not found.
- */
 export const getEntityContext = (
-  entityId: string,
+  stateObj: HassEntity,
   hass: HomeAssistant
 ): EntityContext => {
-  const entity =
-    (hass.entities[entityId] as EntityRegistryDisplayEntry | undefined) || null;
+  const entry = hass.entities[stateObj.entity_id] as
+    | EntityRegistryDisplayEntry
+    | undefined;
 
-  if (!entity) {
+  if (!entry) {
     return {
       entity: null,
       device: null,
@@ -39,10 +32,20 @@ export const getEntityContext = (
       floor: null,
     };
   }
+  return getEntityEntryContext(entry, hass);
+};
 
-  const deviceId = entity?.device_id;
+export const getEntityEntryContext = (
+  entry:
+    | EntityRegistryDisplayEntry
+    | EntityRegistryEntry
+    | ExtEntityRegistryEntry,
+  hass: HomeAssistant
+): EntityContext => {
+  const entity = hass.entities[entry.entity_id];
+  const deviceId = entry?.device_id;
   const device = deviceId ? hass.devices[deviceId] : null;
-  const areaId = entity?.area_id || device?.area_id;
+  const areaId = entry?.area_id || device?.area_id;
   const area = areaId ? hass.areas[areaId] : null;
   const floorId = area?.floor_id;
   const floor = floorId ? hass.floors[floorId] : null;
